% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/stegasaur.R
\name{decode}
\alias{decode}
\title{Decode data from an image}
\usage{
decode(filename)
}
\arguments{
\item{filename}{Filename to load}
}
\description{
Decode data from an image that has been encoded with
\code{\link{encode}}.
}
\seealso{
\code{\link{encode}}, which includes an example.
}

