% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stegasaur.R
\name{encode}
\alias{encode}
\title{Encode data in an image}
\usage{
encode(content, file_in, file_out)
}
\arguments{
\item{content}{Content to encode: any R object is OK (see
\code{\link{lsb_encode}}).}

\item{file_in}{Source filename}

\item{file_out}{Destination filename (can be the same as the
source filename for a destructive update)}
}
\description{
Encode data in an image
}
\examples{
# file included with the package:
file_in <- system.file("man_with_cats.jpg", package = "stegasaur")

# output a png in a temporary directory:
file_out <- tempfile("kitten", fileext = ".png")

# some random data to encode:
x <- runif(10)
encode(x, file_in, file_out)

# extract the data:
decode(file_out)
}
