% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figasaur.R
\name{figasaur}
\alias{figasaur}
\title{Encode plot commands into a plot file}
\usage{
figasaur(code, filename, ...)
}
\arguments{
\item{code}{code use to produce the figure you want}

\item{filename}{name of the file to write}

\item{...}{extra arguments to go to \link{png}}
}
\description{
The code supplied in \code{code} (typially in curly braces) is used to
create a figure which has the plotting code encoded as a character
string in the file. This information can be extracted using
\link{decode}.
}
\examples{
# encode the code in the curly braces into its corresponding plot
path <- tempfile(fileext = ".png")
figasaur({plot(sample(100))}, path)

# extract the information
cat(decode(path))
}
\author{
David L Miller
}
