% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/figuredout.R
\name{figuredout}
\alias{figuredout}
\title{Encode plot commands into a plot file}
\usage{
figuredout(code, filename, ...)
}
\arguments{
\item{code}{code use to produce the figure you want}

\item{filename}{name of the file to write}

\item{\dots}{extra arguments to go to \code{\link{png}}}
}
\description{
The code supplied in \code{code} (in curly braces) is used to create a figure which has the plotting code encoded as a character string in the file. This information can be extracted using \code{\link{decode}}.
}
\examples{
# encode the code in the curly braces into its corresponding plot
# in the file simpleplot.png
figuredout({plot(sample(100))}, "simpleplot.png")
# extract the information
cat(decode("simpleplot.png"))
}
\author{
David L Miller
}

