% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hook_figasaur.R
\name{hook_figasaur}
\alias{hook_figasaur}
\title{'knitr' hook to encode plot code into figures}
\usage{
hook_figasaur(before, options, envir)
}
\arguments{
\item{before, options, envir}{Arguments as required by knitr; please
see references for details}
}
\description{
As plots are generated in a \code{knitr} document, encode the code used
to produce them into the resulting PNG files. This can be done by
setting the chunk option: \code{figasaur = TRUE} after setting
\code{knit_hooks$set(figasaur = hook_figasaur)} at the top of your
\code{knitr} document. The \code{knitr_example.Rmd} file included in this
package shows a simple example.
}
\details{
Note that this only works for PNG files generated for HTML
documents, though it does work when the PNGs are embedded in the
HTML file (e.g. using the \code{self_contained} option in
\link[rmarkdown:render]{rmarkdown::render}. See \link[knitr:opts_hooks]{knitr::opts_hooks} for more information.
}
\examples{
writeLines(readLines(
  system.file("knitr_example.Rmd", package = "stegasaur")))
}
\references{
https://yihui.name/knitr/hooks#chunk_hooks
}
\seealso{
encode
}
\author{
David L Miller
}
