% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/placekitten.R
\name{kitten}
\alias{kitten}
\title{Download a kitten}
\usage{
kitten(width, height, destfile = tempfile("kitten", fileext = ".jpg"), ...)
}
\arguments{
\item{width}{Width, in pixels}

\item{height}{Height, in pixels}

\item{destfile}{File to save into.  By default this is a temporary
file.}

\item{...}{Additional parameters passed through to
\code{\link{download.file}}}
}
\value{
The filename saved in (useful when using the default
\code{destfile})
}
\description{
Download a kitten picture from placekitten.com
}

