% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lsb.R
\name{lsb_encode}
\alias{lsb_decode}
\alias{lsb_encode}
\title{Encode text into a lossless image}
\usage{
lsb_encode(txt, img)

lsb_decode(img)
}
\arguments{
\item{txt}{Text contents to save into the file (raw will be
supported soon).}

\item{img}{An image matrix to save the message into}
}
\description{
Takes an image matrix from something like \code{readPNG}.  Note
that this cannot be saved out as jpeg as the lossy compression
will drop the message
}
\examples{
img <- png::readPNG(system.file("img/Rlogo.png", package="png"))
txt <- "hello from stegasaur"
img2 <- lsb_encode(txt, img)
lsb_decode(img2)
}
\author{
Rich FitzJohn
}

