% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stegasaur.R
\name{decode}
\alias{decode}
\title{Decode data from an image}
\usage{
decode(filename)
}
\arguments{
\item{filename}{Filename to load}
}
\description{
Decode data from an image that has been encoded with
\link{encode}.
}
\seealso{
\link{encode}, which includes an example.
}
