% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsb.R
\name{lsb_encode}
\alias{lsb_encode}
\alias{lsb_decode}
\title{Encode text into a lossless image}
\usage{
lsb_encode(content, img, force_object = FALSE)

lsb_decode(img)
}
\arguments{
\item{content}{Content to save into the image; can be a text
string or an arbitrary R object.}

\item{img}{An image matrix to save the message into}

\item{force_object}{Logical: Force saving a scalar text string as
an R object (will be slightly more space efficient).}
}
\description{
Takes an image matrix from something like \code{readPNG}.  Note
that this cannot be saved out as jpeg as the lossy compression
will drop the message
}
\examples{
img <- png::readPNG(system.file("img/Rlogo.png", package = "png"))
txt <- "hello from stegasaur"
img2 <- lsb_encode(txt, img)
lsb_decode(img2)
}
\author{
Rich FitzJohn
}
